# api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Assisted Migration Agent API
  version: v1
servers:
  - url: /api/v1
paths:
...
  /inspector:
    get:
      summary: Get inspector status
      operationId: getInspectorStatus
      parameters:
        - name: includeVddk
          in: query
          required: false
          description: If true, include uploaded VDDK metadata (`version`, `md5`) when present. omitted if VDDK was never uploaded.
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Inspector status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InspectorStatus'
        '500':
          description: Internal server error
    post:
      summary: Start inspection for VMs
      operationId: startInspection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Start inspection request
              required:
                  - vmIds
                  - credentials
              properties:
                vmIds:
                  type: array
                  items:
                    type: string
                credentials:
                  $ref: '#/components/schemas/VcenterCredentials'
      responses:
        '202':
          description: Inspection started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InspectorStatus'
        '400':
          description: Invalid request
        '409':
          description: Inspector already running
        '500':
          description: Internal server error
    delete:
      summary: Stop inspector entirely
      operationId: stopInspection
      responses:
        '200':
          description: VMs removed from queue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InspectorStatus'
        '400':
          description: Already in canceling state
        '404':
          description: Inspector not running
        '500':
          description: Internal server error
  /inspector/credentials:
    post:
      summary: Validate vCenter credentials for inspector
      operationId: validateInspectorCredentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VcenterCredentials'
      responses:
        "200":
          description: Credentials updated successfully
        "400":
          description: Bad Credentials
        "500":
          description: Internal server error
...
  /inspector/vddk:
    get:
      summary: Get VDDK status
      operationId: getInspectorVddkStatus
      responses:
        '200':
          description: |
            Returns the current properties of the uploaded VMware VDDK tar.
            If the VDDK package has not been uploaded, a 404 response is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VddkProperties'
        '404':
          description: Vddk not found
        '500':
          description: Internal server error
    put:
      summary: Upload VDDK tarball
      operationId: putInspectorVddk
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: VDDK tarball
      responses:
        '200':
          description: Upload successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VddkProperties'
        '409':
          description: Conflict
        '413':
          description: File exceeds 64MB limit
        '400':
          description: Bad request
        '500':
          description: Internal server error
...
components:
  schemas:
    VddkProperties:
      type: object
      required:
        - version
        - md5
      properties:
        version:
          type: string
          description: The matching vSphere Client version
        md5:
          type: string
          description: md5 sum of the uploaded tarball
        bytes:
          type: integer
          format: int64
          description: provided tarball bytes
...
    InspectorStatus:
      type: object
      required:
        - state
      properties:
        state:
          type: string
          enum:
            - ready
            - running
          x-enum-varnames:
            - InspectorStatusStateReady
            - InspectorStatusStateRunning
          description: Inspector state
        vddk:
          allOf:
            - $ref: '#/components/schemas/VddkProperties'
    VcenterCredentials:
      required:
        - url
        - username
        - password
      properties:
        url:
          type: string
          format: uri
          description: vCenter URL
          x-oapi-codegen-extra-tags:
            binding: "required,url"
        username:
          type: string
          minLength: 1
          x-oapi-codegen-extra-tags:
            binding: "required,min=1"
        password:
          type: string
          format: password
          minLength: 1
          x-oapi-codegen-extra-tags:
            binding: "required,min=1"
        cacert:
          type: string
          description: >-
            PEM-encoded CA certificate bundle for verifying the vCenter TLS
            certificate. Mutually exclusive with skipTls.
          x-oapi-codegen-extra-tags:
            binding: "omitempty"
        skipTls:
          type: boolean
          description: >-
            When true, TLS certificate verification is skipped. Must not be set
            to true alongside cacert — that combination returns 400. When
            omitted: if cacert is absent the connection skips TLS verification
            (backwards compatibility); if cacert is provided the connection
            verifies TLS using that certificate — skipTls need not be sent.
...